dispatcher/requirements.txt
Flask~=2.2.2
python-dotenv~=0.20.0
requests~=2.28.1
cloudevents~=1.6.0
